Split tests into separate file
